using dictionary to have better structure for insert to db
