+++ postgres/__init__.py
@@ -1,0 +1,1 @@
+from .postgres import PostGres
